refactor
Split up git-config module into smaller ones by purpose.
