+++ git-config/src/file/git_config/mod.rs
@@ -4,21 +4,13 @@ mod resolve_includes;
 pub use from_env::functions::*;
 pub use resolve_includes::function::resolve_includes;
 
-use bstr::BStr;
-use std::{borrow::Cow, collections::HashMap, convert::TryFrom, fmt::Display, path::Path};
+use std::{borrow::Cow, collections::HashMap, convert::TryFrom};
 
 use crate::{
-    file::{
-        section::{MutableSection, SectionBody},
-        value::{EntryData, MutableMultiValue, MutableValue},
-        Index, Size,
-    },
-    lookup, parser,
-    parser::{
-        parse_from_bytes, parse_from_path, parse_from_str, Error, Event, Key, ParsedSectionHeader, Parser,
-        SectionHeaderName,
-    },
-    value, values, File,
+    file::section::{MutableSection, SectionBody},
+    lookup,
+    parser::{ParsedSectionHeader, SectionHeaderName},
+    File,
 };
 
 /// The section ID is a monotonically increasing ID used to refer to sections.
@@ -33,7 +25,7 @@ use crate::{
 /// words, it's possible that a section may have an ID of 3 but the next section
 /// has an ID of 5.
 #[derive(PartialEq, Eq, Hash, Copy, Clone, PartialOrd, Ord, Debug)]
-pub(crate) struct SectionId(usize);
+pub(crate) struct SectionId(pub(crate) usize);
 
 /// Internal data structure for the section id lookup tree used by
 /// [`GitConfig`]. Note that order in Vec matters as it represents the order
@@ -45,63 +37,9 @@ pub(crate) enum LookupTreeNode<'a> {
     NonTerminal(HashMap<Cow<'a, str>, Vec<SectionId>>),
 }
 
+mod init;
+
 impl<'event> File<'event> {
-    /// Constructs an empty `git-config` file.
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Constructs a `git-config` file from the provided path.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if there was an IO error or if the file wasn't a valid
-    /// git-config file.
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, parser::ParserOrIoError<'static>> {
-        parse_from_path(path).map(Self::from)
-    }
-
-    /// Constructs a `git-config` file from the provided paths in the order provided.
-    /// This is neither zero-copy nor zero-alloc.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if there was an IO error or if a file wasn't a valid
-    /// git-config file.
-    ///
-    /// [`git-config`'s documentation]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-FILES
-    pub fn from_paths(
-        paths: impl IntoIterator<Item = impl AsRef<Path>>,
-        options: from_paths::Options,
-    ) -> Result<Self, from_paths::Error> {
-        let mut target = Self::new();
-        for path in paths {
-            let path = path.as_ref();
-            let mut config = Self::open(path)?;
-            resolve_includes(&mut config, Some(path), options)?;
-            target.append(config);
-        }
-        Ok(target)
-    }
-
-    // TODO: add note indicating that probably a lot if not all information about the original files is currently lost,
-    //       so can't be written back. This will probably change a lot during refactor, so it's not too important now.
-    fn append(&mut self, mut other: Self) {
-        let mut section_indices: Vec<_> = other.section_headers.keys().cloned().collect();
-        // header keys are numeric and ascend in insertion order, hence sorting them gives the order
-        // in which they appear in the config file.
-        section_indices.sort();
-        for section_index in section_indices {
-            let section_header = other.section_headers.remove(&section_index).expect("present");
-            self.push_section(
-                section_header.name.0,
-                section_header.subsection_name,
-                other.sections.remove(&section_index).expect("present"),
-            );
-        }
-    }
-
     /// Returns an interpreted value given a section, an optional subsection and
     /// key.
     ///
@@ -157,66 +95,6 @@ impl<'event> File<'event> {
         key: &str,
     ) -> Option<Result<T, T::Error>> {
         self.raw_value(section_name, subsection_name, key).ok().map(T::try_from)
-    }
-
-    /// Like [`value()`][GitConfig::value()], but returning an `Option` if the string wasn't found.
-    ///
-    /// As strings perform no conversions, this will never fail.
-    pub fn string(
-        &'event self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Option<Cow<'event, BStr>> {
-        self.raw_value(section_name, subsection_name, key)
-            .ok()
-            .map(|v| values::String::from(v).value)
-    }
-
-    /// Like [`value()`][GitConfig::value()], but returning an `Option` if the path wasn't found.
-    ///
-    /// Note that this path is not vetted and should only point to resources which can't be used
-    /// to pose a security risk.
-    ///
-    /// As paths perform no conversions, this will never fail.
-    // TODO: add `secure_path()` or similar to make use of our knowledge of the trust associated with each configuration
-    //       file, maybe even remove the insecure version to force every caller to ask themselves if the resource can
-    //       be used securely or not.
-    pub fn path(
-        &'event self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Option<values::Path<'event>> {
-        self.raw_value(section_name, subsection_name, key)
-            .ok()
-            .map(values::Path::from)
-    }
-
-    /// Like [`value()`][GitConfig::value()], but returning an `Option` if the boolean wasn't found.
-    pub fn boolean(
-        &'event self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Option<Result<bool, value::parse::Error>> {
-        self.raw_value(section_name, subsection_name, key)
-            .ok()
-            .map(|v| values::Boolean::try_from(v).map(|b| b.to_bool()))
-    }
-
-    /// Like [`value()`][GitConfig::value()], but returning an `Option` if the integer wasn't found.
-    pub fn integer(
-        &'event self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Option<Result<i64, value::parse::Error>> {
-        let int = self.raw_value(section_name, subsection_name, key).ok()?;
-        Some(values::Integer::try_from(int.as_ref()).and_then(|b| {
-            b.to_decimal()
-                .ok_or_else(|| value::parse::Error::new("Integer overflow", int.into_owned()))
-        }))
     }
 
     /// Returns all interpreted values given a section, an optional subsection
@@ -577,7 +455,7 @@ impl<'event> File<'event> {
         Ok(())
     }
 
-    /// Returns the number of entries in the config. This ignores any comments.
+    /// Returns the number of values in the config. This ignores any comments.
     /// For example, a config with multiple empty sections will return 0.
     #[must_use]
     pub fn len(&self) -> usize {
@@ -592,893 +470,9 @@ impl<'event> File<'event> {
     }
 }
 
-/// # Raw value API
-///
-/// These functions are the raw value API. Instead of returning Rust structures,
-/// these functions return bytes which may or may not be owned.
-impl<'event> File<'event> {
-    /// Returns an uninterpreted value given a section, an optional subsection
-    /// and key.
-    ///
-    /// Consider [`Self::raw_multi_value`] if you want to get all values of
-    /// a multivar instead.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the key is not in the requested
-    /// section and subsection, or if the section and subsection do not exist.
-    pub fn raw_value<'lookup>(
-        &self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-        key: &'lookup str,
-    ) -> Result<Cow<'_, [u8]>, lookup::existing::Error> {
-        // Note: cannot wrap around the raw_multi_value method because we need
-        // to guarantee that the highest section id is used (so that we follow
-        // the "last one wins" resolution strategy by `git-config`).
-        let key = Key(key.into());
-        for section_id in self
-            .section_ids_by_name_and_subname(section_name, subsection_name)?
-            .iter()
-            .rev()
-        {
-            if let Some(v) = self
-                .sections
-                .get(section_id)
-                .expect("sections does not have section id from section ids")
-                .value(&key)
-            {
-                return Ok(v.to_vec().into());
-            }
-        }
-
-        Err(lookup::existing::Error::KeyMissing)
-    }
-
-    /// Returns a mutable reference to an uninterpreted value given a section,
-    /// an optional subsection and key.
-    ///
-    /// Consider [`Self::raw_multi_value_mut`] if you want to get mutable
-    /// references to all values of a multivar instead.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the key is not in the requested
-    /// section and subsection, or if the section and subsection do not exist.
-    pub fn raw_value_mut<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-        key: &'lookup str,
-    ) -> Result<MutableValue<'_, 'lookup, 'event>, lookup::existing::Error> {
-        let section_ids = self.section_ids_by_name_and_subname(section_name, subsection_name)?;
-        let key = Key(key.into());
-
-        for section_id in section_ids.iter().rev() {
-            let mut size = Size(0);
-            let mut index = Index(0);
-            let mut found_key = false;
-            // todo: iter backwards
-            for (i, event) in self
-                .sections
-                .get(section_id)
-                .expect("sections does not have section id from section ids")
-                .as_ref()
-                .iter()
-                .enumerate()
-            {
-                match event {
-                    Event::Key(event_key) if *event_key == key => {
-                        found_key = true;
-                        size = Size(1);
-                        index = Index(i);
-                    }
-                    Event::Newline(_) | Event::Whitespace(_) | Event::ValueNotDone(_) if found_key => {
-                        size += 1;
-                    }
-                    Event::ValueDone(_) | Event::Value(_) if found_key => {
-                        found_key = false;
-                        size += 1;
-                    }
-                    _ => (),
-                }
-            }
-
-            if size.0 == 0 {
-                continue;
-            }
-
-            return Ok(MutableValue::new(
-                MutableSection::new(
-                    self.sections
-                        .get_mut(section_id)
-                        .expect("sections does not have section id from section ids"),
-                ),
-                key,
-                index,
-                size,
-            ));
-        }
-
-        Err(lookup::existing::Error::KeyMissing)
-    }
-
-    /// Returns all uninterpreted values given a section, an optional subsection
-    /// and key.
-    ///
-    /// # Examples
-    ///
-    /// If you have the following config:
-    ///
-    /// ```text
-    /// [core]
-    ///     a = b
-    /// [core]
-    ///     a = c
-    ///     a = d
-    /// ```
-    ///
-    /// Attempting to get all values of `a` yields the following:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// # let git_config = git_config::File::try_from("[core]a=b\n[core]\na=c\na=d").unwrap();
-    /// assert_eq!(
-    ///     git_config.raw_multi_value("core", None, "a").unwrap(),
-    ///     vec![
-    ///         Cow::<[u8]>::Borrowed(b"b"),
-    ///         Cow::<[u8]>::Borrowed(b"c"),
-    ///         Cow::<[u8]>::Borrowed(b"d"),
-    ///     ],
-    /// );
-    /// ```
-    ///
-    /// Consider [`Self::raw_value`] if you want to get the resolved single
-    /// value for a given key, if your key does not support multi-valued values.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the key is not in any requested
-    /// section and subsection, or if no instance of the section and subsections
-    /// exist.
-    pub fn raw_multi_value(
-        &self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Result<Vec<Cow<'_, [u8]>>, lookup::existing::Error> {
-        let mut values = vec![];
-        for section_id in self.section_ids_by_name_and_subname(section_name, subsection_name)? {
-            values.extend(
-                self.sections
-                    .get(&section_id)
-                    .expect("sections does not have section id from section ids")
-                    .values(&Key(key.into()))
-                    .iter()
-                    .map(|v| Cow::Owned(v.to_vec())),
-            );
-        }
-
-        if values.is_empty() {
-            Err(lookup::existing::Error::KeyMissing)
-        } else {
-            Ok(values)
-        }
-    }
-
-    /// Similar to [`multi_value(…)`][GitConfig::multi_value()] but returning strings if at least one of them was found.
-    pub fn strings(&self, section_name: &str, subsection_name: Option<&str>, key: &str) -> Option<Vec<Cow<'_, BStr>>> {
-        self.raw_multi_value(section_name, subsection_name, key)
-            .ok()
-            .map(|values| values.into_iter().map(|v| values::String::from(v).value).collect())
-    }
-
-    /// Similar to [`multi_value(…)`][GitConfig::multi_value()] but returning integers if at least one of them was found
-    /// and if none of them overflows.
-    pub fn integers(
-        &self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Option<Result<Vec<i64>, value::parse::Error>> {
-        self.raw_multi_value(section_name, subsection_name, key)
-            .ok()
-            .map(|values| {
-                values
-                    .into_iter()
-                    .map(|v| {
-                        values::Integer::try_from(v.as_ref()).and_then(|int| {
-                            int.to_decimal()
-                                .ok_or_else(|| value::parse::Error::new("Integer overflow", v.into_owned()))
-                        })
-                    })
-                    .collect()
-            })
-    }
-
-    /// Returns mutable references to all uninterpreted values given a section,
-    /// an optional subsection and key.
-    ///
-    /// # Examples
-    ///
-    /// If you have the following config:
-    ///
-    /// ```text
-    /// [core]
-    ///     a = b
-    /// [core]
-    ///     a = c
-    ///     a = d
-    /// ```
-    ///
-    /// Attempting to get all values of `a` yields the following:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// # let mut git_config = git_config::File::try_from("[core]a=b\n[core]\na=c\na=d").unwrap();
-    /// assert_eq!(
-    ///     git_config.raw_multi_value("core", None, "a")?,
-    ///     vec![
-    ///         Cow::Borrowed(b"b"),
-    ///         Cow::Borrowed(b"c"),
-    ///         Cow::Borrowed(b"d")
-    ///     ]
-    /// );
-    ///
-    /// git_config.raw_multi_value_mut("core", None, "a")?.set_str_all("g");
-    ///
-    /// assert_eq!(
-    ///     git_config.raw_multi_value("core", None, "a")?,
-    ///     vec![
-    ///         Cow::Borrowed(b"g"),
-    ///         Cow::Borrowed(b"g"),
-    ///         Cow::Borrowed(b"g")
-    ///     ],
-    /// );
-    /// # Ok::<(), git_config::lookup::existing::Error>(())
-    /// ```
-    ///
-    /// Consider [`Self::raw_value`] if you want to get the resolved single
-    /// value for a given key, if your key does not support multi-valued values.
-    ///
-    /// Note that this operation is relatively expensive, requiring a full
-    /// traversal of the config.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the key is not in any requested
-    /// section and subsection, or if no instance of the section and subsections
-    /// exist.
-    pub fn raw_multi_value_mut<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-        key: &'lookup str,
-    ) -> Result<MutableMultiValue<'_, 'lookup, 'event>, lookup::existing::Error> {
-        let section_ids = self.section_ids_by_name_and_subname(section_name, subsection_name)?;
-        let key = Key(key.into());
-
-        let mut offsets = HashMap::new();
-        let mut entries = vec![];
-        for section_id in section_ids.iter().rev() {
-            let mut last_boundary = 0;
-            let mut found_key = false;
-            let mut offset_list = vec![];
-            let mut offset_index = 0;
-            for (i, event) in self
-                .sections
-                .get(section_id)
-                .expect("sections does not have section id from section ids")
-                .as_ref()
-                .iter()
-                .enumerate()
-            {
-                match event {
-                    Event::Key(event_key) if *event_key == key => {
-                        found_key = true;
-                        offset_list.push(i - last_boundary);
-                        offset_index += 1;
-                        last_boundary = i;
-                    }
-                    Event::Value(_) | Event::ValueDone(_) if found_key => {
-                        found_key = false;
-                        entries.push(EntryData::new(*section_id, offset_index));
-                        offset_list.push(i - last_boundary + 1);
-                        offset_index += 1;
-                        last_boundary = i + 1;
-                    }
-                    _ => (),
-                }
-            }
-            offsets.insert(*section_id, offset_list);
-        }
-
-        entries.sort();
-
-        if entries.is_empty() {
-            Err(lookup::existing::Error::KeyMissing)
-        } else {
-            Ok(MutableMultiValue::new(&mut self.sections, key, entries, offsets))
-        }
-    }
-
-    /// Sets a value in a given section, optional subsection, and key value.
-    ///
-    /// # Examples
-    ///
-    /// Given the config,
-    ///
-    /// ```text
-    /// [core]
-    ///     a = b
-    /// [core]
-    ///     a = c
-    ///     a = d
-    /// ```
-    ///
-    /// Setting a new value to the key `core.a` will yield the following:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// # let mut git_config = git_config::File::try_from("[core]a=b\n[core]\na=c\na=d").unwrap();
-    /// git_config.set_raw_value("core", None, "a", vec![b'e'])?;
-    /// assert_eq!(git_config.raw_value("core", None, "a")?, Cow::Borrowed(b"e"));
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// This errors if any lookup input (section, subsection, and key value) fails.
-    pub fn set_raw_value<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-        key: &'lookup str,
-        new_value: Vec<u8>,
-    ) -> Result<(), lookup::existing::Error> {
-        self.raw_value_mut(section_name, subsection_name, key)
-            .map(|mut entry| entry.set_bytes(new_value))
-    }
-
-    /// Sets a multivar in a given section, optional subsection, and key value.
-    ///
-    /// This internally zips together the new values and the existing values.
-    /// As a result, if more new values are provided than the current amount of
-    /// multivars, then the latter values are not applied. If there are less
-    /// new values than old ones then the remaining old values are unmodified.
-    ///
-    /// **Note**: Mutation order is _not_ guaranteed and is non-deterministic.
-    /// If you need finer control over which values of the multivar are set,
-    /// consider using [`raw_multi_value_mut`], which will let you iterate
-    /// and check over the values instead. This is best used as a convenience
-    /// function for setting multivars whose values should be treated as an
-    /// unordered set.
-    ///
-    /// # Examples
-    ///
-    /// Let us use the follow config for all examples:
-    ///
-    /// ```text
-    /// [core]
-    ///     a = b
-    /// [core]
-    ///     a = c
-    ///     a = d
-    /// ```
-    ///
-    /// Setting an equal number of values:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// # let mut git_config = git_config::File::try_from("[core]a=b\n[core]\na=c\na=d").unwrap();
-    /// let new_values: Vec<Cow<'_, [u8]>> = vec![
-    ///     Cow::Borrowed(b"x"),
-    ///     Cow::Borrowed(b"y"),
-    ///     Cow::Borrowed(b"z"),
-    /// ];
-    /// git_config.set_raw_multi_value("core", None, "a", new_values.into_iter())?;
-    /// let fetched_config = git_config.raw_multi_value("core", None, "a")?;
-    /// assert!(fetched_config.contains(&Cow::Borrowed(b"x")));
-    /// assert!(fetched_config.contains(&Cow::Borrowed(b"y")));
-    /// assert!(fetched_config.contains(&Cow::Borrowed(b"z")));
-    /// # Ok::<(), git_config::lookup::existing::Error>(())
-    /// ```
-    ///
-    /// Setting less than the number of present values sets the first ones found:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// # let mut git_config = git_config::File::try_from("[core]a=b\n[core]\na=c\na=d").unwrap();
-    /// let new_values: Vec<Cow<'_, [u8]>> = vec![
-    ///     Cow::Borrowed(b"x"),
-    ///     Cow::Borrowed(b"y"),
-    /// ];
-    /// git_config.set_raw_multi_value("core", None, "a", new_values.into_iter())?;
-    /// let fetched_config = git_config.raw_multi_value("core", None, "a")?;
-    /// assert!(fetched_config.contains(&Cow::Borrowed(b"x")));
-    /// assert!(fetched_config.contains(&Cow::Borrowed(b"y")));
-    /// # Ok::<(), git_config::lookup::existing::Error>(())
-    /// ```
-    ///
-    /// Setting more than the number of present values discards the rest:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// # let mut git_config = git_config::File::try_from("[core]a=b\n[core]\na=c\na=d").unwrap();
-    /// let new_values: Vec<Cow<'_, [u8]>> = vec![
-    ///     Cow::Borrowed(b"x"),
-    ///     Cow::Borrowed(b"y"),
-    ///     Cow::Borrowed(b"z"),
-    ///     Cow::Borrowed(b"discarded"),
-    /// ];
-    /// git_config.set_raw_multi_value("core", None, "a", new_values.into_iter())?;
-    /// assert!(!git_config.raw_multi_value("core", None, "a")?.contains(&Cow::Borrowed(b"discarded")));
-    /// # Ok::<(), git_config::lookup::existing::Error>(())
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// This errors if any lookup input (section, subsection, and key value) fails.
-    ///
-    /// [`raw_multi_value_mut`]: Self::raw_multi_value_mut
-    pub fn set_raw_multi_value<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-        key: &'lookup str,
-        new_values: impl Iterator<Item = Cow<'event, [u8]>>,
-    ) -> Result<(), lookup::existing::Error> {
-        self.raw_multi_value_mut(section_name, subsection_name, key)
-            .map(|mut v| v.set_values(new_values))
-    }
-}
-
-/// Private helper functions
-impl<'event> File<'event> {
-    /// Adds a new section to the config file.
-    fn push_section_internal(
-        &mut self,
-        // current_section_name: Option<SectionHeaderName<'event>>,
-        // current_subsection_name: Option<Cow<'event, str>>,
-        header: ParsedSectionHeader<'event>,
-        section: SectionBody<'event>,
-    ) -> MutableSection<'_, 'event> {
-        let new_section_id = SectionId(self.section_id_counter);
-        self.section_headers.insert(new_section_id, header.clone());
-        self.sections.insert(new_section_id, section);
-        let lookup = self.section_lookup_tree.entry(header.name).or_default();
-
-        let mut found_node = false;
-        if let Some(subsection_name) = header.subsection_name {
-            for node in lookup.iter_mut() {
-                if let LookupTreeNode::NonTerminal(subsection) = node {
-                    found_node = true;
-                    subsection
-                        // Clones the cow, not the inner borrowed str.
-                        .entry(subsection_name.clone())
-                        .or_default()
-                        .push(new_section_id);
-                    break;
-                }
-            }
-            if !found_node {
-                let mut map = HashMap::new();
-                map.insert(subsection_name, vec![new_section_id]);
-                lookup.push(LookupTreeNode::NonTerminal(map));
-            }
-        } else {
-            for node in lookup.iter_mut() {
-                if let LookupTreeNode::Terminal(vec) = node {
-                    found_node = true;
-                    vec.push(new_section_id);
-                    break;
-                }
-            }
-            if !found_node {
-                lookup.push(LookupTreeNode::Terminal(vec![new_section_id]));
-            }
-        }
-        self.section_order.push_back(new_section_id);
-        self.section_id_counter += 1;
-        self.sections.get_mut(&new_section_id).map(MutableSection::new).unwrap()
-    }
-
-    /// Returns the mapping between section and subsection name to section ids.
-    fn section_ids_by_name_and_subname<'lookup>(
-        &self,
-        section_name: impl Into<SectionHeaderName<'lookup>>,
-        subsection_name: Option<&'lookup str>,
-    ) -> Result<Vec<SectionId>, lookup::existing::Error> {
-        let section_name = section_name.into();
-        let section_ids = self
-            .section_lookup_tree
-            .get(&section_name)
-            .ok_or(lookup::existing::Error::SectionMissing)?;
-        let mut maybe_ids = None;
-        // Don't simplify if and matches here -- the for loop currently needs
-        // `n + 1` checks, while the if and matches will result in the for loop
-        // needing `2n` checks.
-        if let Some(subsection_name) = subsection_name {
-            for node in section_ids {
-                if let LookupTreeNode::NonTerminal(subsection_lookup) = node {
-                    maybe_ids = subsection_lookup.get(subsection_name);
-                    break;
-                }
-            }
-        } else {
-            for node in section_ids {
-                if let LookupTreeNode::Terminal(subsection_lookup) = node {
-                    maybe_ids = Some(subsection_lookup);
-                    break;
-                }
-            }
-        }
-        maybe_ids
-            .map(Vec::to_owned)
-            .ok_or(lookup::existing::Error::SubSectionMissing)
-    }
-
-    fn section_ids_by_name<'lookup>(
-        &self,
-        section_name: impl Into<SectionHeaderName<'lookup>>,
-    ) -> Result<Vec<SectionId>, lookup::existing::Error> {
-        let section_name = section_name.into();
-        self.section_lookup_tree
-            .get(&section_name)
-            .map(|lookup| {
-                lookup
-                    .iter()
-                    .flat_map(|node| match node {
-                        LookupTreeNode::Terminal(v) => v.clone(),
-                        LookupTreeNode::NonTerminal(v) => v.values().flatten().copied().collect(),
-                    })
-                    .collect()
-            })
-            .ok_or(lookup::existing::Error::SectionMissing)
-    }
-}
-
-impl<'a> TryFrom<&'a str> for File<'a> {
-    type Error = Error<'a>;
-
-    /// Convenience constructor. Attempts to parse the provided string into a
-    /// [`GitConfig`]. See [`parse_from_str`] for more information.
-    ///
-    /// [`parse_from_str`]: crate::parser::parse_from_str
-    fn try_from(s: &'a str) -> Result<File<'a>, Self::Error> {
-        parse_from_str(s).map(Self::from)
-    }
-}
-
-impl<'a> TryFrom<&'a [u8]> for File<'a> {
-    type Error = Error<'a>;
-
-    /// Convenience constructor. Attempts to parse the provided byte string into
-    //// a [`GitConfig`]. See [`parse_from_bytes`] for more information.
-    ///
-    /// [`parse_from_bytes`]: crate::parser::parse_from_bytes
-    fn try_from(value: &'a [u8]) -> Result<File<'a>, Self::Error> {
-        parse_from_bytes(value).map(File::from)
-    }
-}
-
-impl<'a> TryFrom<&'a Vec<u8>> for File<'a> {
-    type Error = Error<'a>;
-
-    /// Convenience constructor. Attempts to parse the provided byte string into
-    //// a [`GitConfig`]. See [`parse_from_bytes`] for more information.
-    ///
-    /// [`parse_from_bytes`]: crate::parser::parse_from_bytes
-    fn try_from(value: &'a Vec<u8>) -> Result<File<'a>, Self::Error> {
-        parse_from_bytes(value).map(File::from)
-    }
-}
-
-impl<'a> From<Parser<'a>> for File<'a> {
-    fn from(parser: Parser<'a>) -> Self {
-        let mut new_self = Self::default();
-
-        // Current section that we're building
-        let mut prev_section_header = None;
-        let mut section_events = SectionBody::new();
-
-        #[allow(clippy::explicit_into_iter_loop)] // it's not really an iterator (yet), needs streaming iterator support
-        for event in parser.into_iter() {
-            match event {
-                Event::SectionHeader(header) => {
-                    if let Some(prev_header) = prev_section_header.take() {
-                        new_self.push_section_internal(prev_header, section_events);
-                    } else {
-                        new_self.frontmatter_events = section_events;
-                    }
-                    prev_section_header = Some(header);
-                    section_events = SectionBody::new();
-                }
-                e @ Event::Key(_)
-                | e @ Event::Value(_)
-                | e @ Event::ValueNotDone(_)
-                | e @ Event::ValueDone(_)
-                | e @ Event::KeyValueSeparator => section_events.as_mut().push(e),
-                e @ Event::Comment(_) | e @ Event::Newline(_) | e @ Event::Whitespace(_) => {
-                    section_events.as_mut().push(e);
-                }
-            }
-        }
-
-        // The last section doesn't get pushed since we only push if there's a
-        // new section header, so we need to call push one more time.
-        if let Some(header) = prev_section_header {
-            new_self.push_section_internal(header, section_events);
-        } else {
-            new_self.frontmatter_events = section_events;
-        }
-
-        new_self
-    }
-}
-
-impl From<File<'_>> for Vec<u8> {
-    fn from(c: File) -> Self {
-        c.into()
-    }
-}
-
-impl From<&File<'_>> for Vec<u8> {
-    fn from(config: &File) -> Self {
-        let mut value = Self::new();
-
-        for events in config.frontmatter_events.as_ref() {
-            value.extend(events.to_vec());
-        }
-
-        for section_id in &config.section_order {
-            value.extend(
-                config
-                    .section_headers
-                    .get(section_id)
-                    .expect("section_header does not contain section id from section_order")
-                    .to_vec(),
-            );
-
-            for event in config
-                .sections
-                .get(section_id)
-                .expect("sections does not contain section id from section_order")
-                .as_ref()
-            {
-                value.extend(event.to_vec());
-            }
-        }
-
-        value
-    }
-}
-
-impl Display for File<'_> {
-    /// Note that this is a best-effort attempt at printing a `GitConfig`. If
-    /// there are non UTF-8 values in your config, this will _NOT_ render as
-    /// read.
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for front_matter in self.frontmatter_events.as_ref() {
-            front_matter.fmt(f)?;
-        }
-
-        for section_id in &self.section_order {
-            self.section_headers.get(section_id).unwrap().fmt(f)?;
-            for event in self.sections.get(section_id).unwrap().as_ref() {
-                event.fmt(f)?;
-            }
-        }
-
-        Ok(())
-    }
-}
+mod access;
+mod impls;
+mod utils;
 
 #[cfg(test)]
-mod from_parser {
-    use super::{Cow, Event, File, HashMap, LookupTreeNode, SectionBody, SectionId, TryFrom};
-    use crate::{
-        parser::SectionHeaderName,
-        test_util::{name_event, newline_event, section_header, value_event},
-    };
-
-    #[test]
-    fn parse_empty() {
-        let config = File::try_from("").unwrap();
-        assert!(config.section_headers.is_empty());
-        assert_eq!(config.section_id_counter, 0);
-        assert!(config.section_lookup_tree.is_empty());
-        assert!(config.sections.is_empty());
-        assert!(config.section_order.is_empty());
-    }
-
-    #[test]
-    fn parse_single_section() {
-        let mut config = File::try_from("[core]\na=b\nc=d").unwrap();
-        let expected_separators = {
-            let mut map = HashMap::new();
-            map.insert(SectionId(0), section_header("core", None));
-            map
-        };
-        assert_eq!(config.section_headers, expected_separators);
-        assert_eq!(config.section_id_counter, 1);
-        let expected_lookup_tree = {
-            let mut tree = HashMap::new();
-            tree.insert(
-                SectionHeaderName(Cow::Borrowed("core")),
-                vec![LookupTreeNode::Terminal(vec![SectionId(0)])],
-            );
-            tree
-        };
-        assert_eq!(config.section_lookup_tree, expected_lookup_tree);
-        let expected_sections = {
-            let mut sections = HashMap::new();
-            sections.insert(
-                SectionId(0),
-                SectionBody::from(vec![
-                    newline_event(),
-                    name_event("a"),
-                    Event::KeyValueSeparator,
-                    value_event("b"),
-                    newline_event(),
-                    name_event("c"),
-                    Event::KeyValueSeparator,
-                    value_event("d"),
-                ]),
-            );
-            sections
-        };
-        assert_eq!(config.sections, expected_sections);
-        assert_eq!(config.section_order.make_contiguous(), &[SectionId(0)]);
-    }
-
-    #[test]
-    fn parse_single_subsection() {
-        let mut config = File::try_from("[core.sub]\na=b\nc=d").unwrap();
-        let expected_separators = {
-            let mut map = HashMap::new();
-            map.insert(SectionId(0), section_header("core", (".", "sub")));
-            map
-        };
-        assert_eq!(config.section_headers, expected_separators);
-        assert_eq!(config.section_id_counter, 1);
-        let expected_lookup_tree = {
-            let mut tree = HashMap::new();
-            let mut inner_tree = HashMap::new();
-            inner_tree.insert(Cow::Borrowed("sub"), vec![SectionId(0)]);
-            tree.insert(
-                SectionHeaderName(Cow::Borrowed("core")),
-                vec![LookupTreeNode::NonTerminal(inner_tree)],
-            );
-            tree
-        };
-        assert_eq!(config.section_lookup_tree, expected_lookup_tree);
-        let expected_sections = {
-            let mut sections = HashMap::new();
-            sections.insert(
-                SectionId(0),
-                SectionBody::from(vec![
-                    newline_event(),
-                    name_event("a"),
-                    Event::KeyValueSeparator,
-                    value_event("b"),
-                    newline_event(),
-                    name_event("c"),
-                    Event::KeyValueSeparator,
-                    value_event("d"),
-                ]),
-            );
-            sections
-        };
-        assert_eq!(config.sections, expected_sections);
-        assert_eq!(config.section_order.make_contiguous(), &[SectionId(0)]);
-    }
-
-    #[test]
-    fn parse_multiple_sections() {
-        let mut config = File::try_from("[core]\na=b\nc=d\n[other]e=f").unwrap();
-        let expected_separators = {
-            let mut map = HashMap::new();
-            map.insert(SectionId(0), section_header("core", None));
-            map.insert(SectionId(1), section_header("other", None));
-            map
-        };
-        assert_eq!(config.section_headers, expected_separators);
-        assert_eq!(config.section_id_counter, 2);
-        let expected_lookup_tree = {
-            let mut tree = HashMap::new();
-            tree.insert(
-                SectionHeaderName(Cow::Borrowed("core")),
-                vec![LookupTreeNode::Terminal(vec![SectionId(0)])],
-            );
-            tree.insert(
-                SectionHeaderName(Cow::Borrowed("other")),
-                vec![LookupTreeNode::Terminal(vec![SectionId(1)])],
-            );
-            tree
-        };
-        assert_eq!(config.section_lookup_tree, expected_lookup_tree);
-        let expected_sections = {
-            let mut sections = HashMap::new();
-            sections.insert(
-                SectionId(0),
-                SectionBody::from(vec![
-                    newline_event(),
-                    name_event("a"),
-                    Event::KeyValueSeparator,
-                    value_event("b"),
-                    newline_event(),
-                    name_event("c"),
-                    Event::KeyValueSeparator,
-                    value_event("d"),
-                    newline_event(),
-                ]),
-            );
-            sections.insert(
-                SectionId(1),
-                SectionBody::from(vec![name_event("e"), Event::KeyValueSeparator, value_event("f")]),
-            );
-            sections
-        };
-        assert_eq!(config.sections, expected_sections);
-        assert_eq!(config.section_order.make_contiguous(), &[SectionId(0), SectionId(1)]);
-    }
-
-    #[test]
-    fn parse_multiple_duplicate_sections() {
-        let mut config = File::try_from("[core]\na=b\nc=d\n[core]e=f").unwrap();
-        let expected_separators = {
-            let mut map = HashMap::new();
-            map.insert(SectionId(0), section_header("core", None));
-            map.insert(SectionId(1), section_header("core", None));
-            map
-        };
-        assert_eq!(config.section_headers, expected_separators);
-        assert_eq!(config.section_id_counter, 2);
-        let expected_lookup_tree = {
-            let mut tree = HashMap::new();
-            tree.insert(
-                SectionHeaderName(Cow::Borrowed("core")),
-                vec![LookupTreeNode::Terminal(vec![SectionId(0), SectionId(1)])],
-            );
-            tree
-        };
-        assert_eq!(config.section_lookup_tree, expected_lookup_tree);
-        let expected_sections = {
-            let mut sections = HashMap::new();
-            sections.insert(
-                SectionId(0),
-                SectionBody::from(vec![
-                    newline_event(),
-                    name_event("a"),
-                    Event::KeyValueSeparator,
-                    value_event("b"),
-                    newline_event(),
-                    name_event("c"),
-                    Event::KeyValueSeparator,
-                    value_event("d"),
-                    newline_event(),
-                ]),
-            );
-            sections.insert(
-                SectionId(1),
-                SectionBody::from(vec![name_event("e"), Event::KeyValueSeparator, value_event("f")]),
-            );
-            sections
-        };
-        assert_eq!(config.sections, expected_sections);
-        assert_eq!(config.section_order.make_contiguous(), &[SectionId(0), SectionId(1)]);
-    }
-}
+mod try_from_str_tests;
